Move numba skip helper function
